+++ apps/docs/components/mdx-content.tsx
@@ -2,7 +2,7 @@
 "use client";
 import type {MDXComponents as MDXComponentsType} from "mdx/types";
 
-import {useMDXComponent} from "next-contentlayer/hooks";
+import {useMDXComponent} from "next-contentlayer2/hooks";
 
 import {MDXComponents} from "./mdx-components";
