+++ apps/docs/contentlayer.config.js
@@ -1,10 +1,11 @@
-import {defineDocumentType, defineNestedType, makeSource} from "contentlayer/source-files";
+import {defineDocumentType, defineNestedType, makeSource} from "contentlayer2/source-files";
 import remarkGfm from "remark-gfm";
 import rehypeSlug from "rehype-slug";
 import {visit} from "unist-util-visit";
+import RawPlugin from 'esbuild-plugin-raw'
 import pluginCodeBlock from "./plugins/codeBlock";
 
-/** @type {import('contentlayer/source-files').ComputedFields} */
+/** @type {import('contentlayer2/source-files').ComputedFields} */
 const computedFields = {
   slug: {
     type: "string",
@@ -15,8 +16,7 @@ const computedFields = {
     resolve: (doc) => doc._raw.flattenedPath.split("/").slice(1).join("/"),
   },
   url: {type: "string", resolve: (doc) => `/${doc._raw.flattenedPath}`},
-}
-
+};
 
 export const Doc = defineDocumentType(() => ({
   name: "Doc",
@@ -27,9 +27,8 @@ export const Doc = defineDocumentType(() => ({
     description: {type: "string", required: false},
     date: {type: "date", required: false},
   },
-  computedFields
+  computedFields,
 }));
-
 
 const AuthorProperties = defineNestedType(() => ({
   name: "AuthorProperties",
@@ -38,7 +37,7 @@ const AuthorProperties = defineNestedType(() => ({
     link: {type: "string", required: false},
     avatar: {type: "string", required: false},
     username: {type: "string", required: false},
-  }
+  },
 }));
 
 export const BlogPost = defineDocumentType(() => ({
@@ -50,8 +49,8 @@ export const BlogPost = defineDocumentType(() => ({
     description: {type: "string", required: true},
     date: {type: "date", required: true},
     draft: {type: "boolean", required: false},
-    tags: { type: 'list', of: { type: 'string' } },
-    author: {type: "nested",of: AuthorProperties, required: false},
+    tags: {type: "list", of: {type: "string"}},
+    author: {type: "nested", of: AuthorProperties, required: false},
     image: {type: "string", required: false},
   },
   computedFields: {
@@ -63,7 +62,7 @@ export const BlogPost = defineDocumentType(() => ({
         const date = new Date(doc.date);
         const options = {year: "numeric", month: "long", day: "numeric"};
         return date.toLocaleDateString("en-US", options);
-      }
+      },
     },
     // add https://nextui.org to the image path
     imageAsParams: {
@@ -73,9 +72,9 @@ export const BlogPost = defineDocumentType(() => ({
         if (image) {
           return `https://nextui.org${image}`;
         }
-      }
-    }
-  }
+      },
+    },
+  },
 }));
 
 export default makeSource({
@@ -83,6 +82,12 @@ export default makeSource({
   documentTypes: [Doc, BlogPost],
   mdx: {
     remarkPlugins: [remarkGfm, pluginCodeBlock],
+    esbuildOptions(options) {
+      options.plugins ||= [];
+      options.plugins.unshift(RawPlugin());
+ 
+      return options;
+    },
     rehypePlugins: [
       rehypeSlug,
       () => (tree) => {
