+++ apps/docs/global.d.ts
@@ -1,0 +1,4 @@
+declare module "*.tsx?raw" {
+  const content: string;
+  export default content;
+}
