+++ apps/docs/libs/rehype-highlight-word.js
@@ -1,13 +1,13 @@
-const hastToHtml = require("hast-util-to-html");
-const unified = require("unified");
-const parse = require("rehype-parse");
+import {unified} from "unified";
+import {toHtml} from "hast-util-to-html";
+import rehypeParse from "rehype-parse";
 
 const CALLOUT = /__(.*?)__/g;
 
-module.exports = (code) => {
-  const html = hastToHtml(code);
+export default function (code) {
+  const html = toHtml(code);
   const result = html.replace(CALLOUT, (_, text) => `<span class="highlight-word">${text}</span>`);
-  const hast = unified().use(parse, {emitParseErrors: true, fragment: true}).parse(result);
+  const hast = unified().use(rehypeParse, {emitParseErrors: true, fragment: true}).parse(result);
 
   return hast.children;
-};
+}
