+++ apps/docs/package.json
@@ -3,12 +3,12 @@
   "version": "2.0.0",
   "private": true,
   "scripts": {
-    "dev": "rimraf .contentlayer && concurrently \"contentlayer dev\" \"next dev\"",
-    "build": "contentlayer build && next build",
+    "dev": "rimraf .contentlayer && concurrently \"contentlayer2 dev\" \"next dev\"",
+    "build": "contentlayer2 build && next build",
     "build:analyze": "ANALYZE=true next build",
     "start": "next start",
     "lint": "next lint",
-    "typecheck": "contentlayer build && tsc --noEmit",
+    "typecheck": "contentlayer2 build && tsc --noEmit",
     "preinstall": "node preinstall.js",
     "build:sponsors": "tsx scripts/build-sponsors.ts",
     "update:search-meta": "tsx scripts/update-search-meta.ts",
@@ -53,19 +53,19 @@
     "canvas-confetti": "^1.9.2",
     "cmdk": "^0.2.0",
     "color2k": "^2.0.2",
-    "contentlayer": "^0.3.4",
+    "contentlayer2": "^0.5.1",
     "date-fns": "^2.30.0",
     "framer-motion": "^11.1.7",
     "github-slugger": "^2.0.0",
     "gray-matter": "^4.0.3",
-    "hast-util-to-html": "7.1.2",
+    "hast-util-to-html": "9.0.3",
     "lodash": "^4.17.21",
     "marked": "^5.1.0",
     "match-sorter": "^6.3.1",
     "mini-svg-data-uri": "^1.4.3",
     "mitt": "3.0.0",
     "next": "13.5.1",
-    "next-contentlayer": "^0.3.4",
+    "next-contentlayer2": "^0.5.1",
     "next-themes": "^0.2.1",
     "nprogress": "^0.2.0",
     "parse-numeric-range": "1.2.0",
@@ -78,22 +78,20 @@
     "react-multi-ref": "^1.0.1",
     "react-wrap-balancer": "^1.0.0",
     "refractor": "3.3.1",
-    "rehype": "11.0.0",
-    "rehype-parse": "7.0.1",
-    "rehype-pretty-code": "^0.10.0",
-    "rehype-slug": "^5.1.0",
-    "remark": "^14.0.3",
-    "remark-autolink-headings": "^6.0.1",
-    "remark-gfm": "^3.0.1",
-    "remark-slug": "^6.0.0",
+    "rehype": "13.0.2",
+    "rehype-parse": "9.0.1",
+    "rehype-pretty-code": "^0.14.0",
+    "rehype-slug": "^6.0.0",
+    "remark": "^15.0.1",
+    "remark-gfm": "^4.0.0",
     "rss": "^1.2.2",
     "scroll-into-view-if-needed": "3.0.10",
     "sharp": "^0.32.1",
     "shelljs": "^0.8.4",
     "swr": "^2.2.1",
     "tailwind-variants": "^0.1.20",
-    "unified": "^9.2.2",
-    "unist-util-visit": "^4.1.2",
+    "unified": "^11.0.5",
+    "unist-util-visit": "^5.0.0",
     "usehooks-ts": "^2.9.1",
     "zustand": "^4.3.8"
   },
@@ -119,6 +117,7 @@
     "algoliasearch": "^4.10.3",
     "autoprefixer": "^10.4.14",
     "dotenv": "^16.0.1",
+    "esbuild-plugin-raw": "0.1.8",
     "eslint-config-next": "^13.5.1",
     "markdown-toc": "^1.2.0",
     "next-sitemap": "^4.1.8",
@@ -127,7 +126,7 @@
     "prettier": "^2.7.1",
     "tailwindcss": "^3.4.0",
     "tsx": "^3.8.2",
-    "typescript": "^4.9.5",
+    "typescript": "^5.5.0",
     "uuid": "^8.3.2"
   }
 }
